1222 아주이상하네요 삭제
