+++ src/main/java/com/kh/bigFish/shop/model/service/ShopService.java
@@ -26,4 +26,7 @@ public interface ShopService {
 
 	int deleteShop(int sno);
 	
+	int insertParameter(String parameter);
+	
+	int updatePg_token(String j);
 }
