+++ src/main/java/com/kh/bigFish/store/model/service/StoreServiceImpl.java
@@ -219,6 +219,7 @@ public class StoreServiceImpl implements StoreService{
 	@Override
 	public Store getStoreNameInfoForChat(Store s) {
 		return storeDao.getStoreNameInfoForChat(sqlSession, s);
+	}
 
 	public List<String> fishKindList(String City1, String City2, String City3, String City4, String City5, String City6) {
 		return storeDao.fishKindList(sqlSession, City1, City2, City3, City4, City5, City6);
