+++ src/main/resources/mappers/kakaoRequestDto-mapper.xml
@@ -2,23 +2,59 @@
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
  "https:/mybatis.org/dtd/mybatis-3-mapper.dtd" >
 
-<mapper namespace="kakaoRequestDto-mapper">
+<mapper namespace="kakaoRequestDtoMapper">
 	<resultMap type="KakaoRequestDto" id="kakaoRequestDtoResult">
-		<result column="C_NO" property="payNo"/>
-		<result column="C_ID" property="cid"/>
-		<result column="PARTNER_ORDER_ID" property="partner_order_id"/>
-		<result column="PARTNER_USER_ID" property="partner_user_id"/>
-		<result column="ITEM_NAME" property="item_name"/>
-		<result column="QUANTITY" property="quantity"/>
-		<result column="TOTAL_AMOUNT" property="total_amount"/>
-		<result column="VAT_AMOUNT" property="vat_amount"/>
-		<result column="TAX_FREE_AMOUNT" property="tax_free_amount"/>
-		<result column="APPROVAL_URL" property="approval_url"/>
-		<result column="FAIL_URL" property="fail_url"/>
-		<result column="CANCEL_URL" property="cancel_url"/>
-		<result column="PG_TOKEN" property="pg_token"/>
-		<result column="PRODUCT_NO" property="productNo"/>
+		<result column="C_NO" property="payNo" />
+		<result column="C_ID" property="cid" />
+		<result column="PARTNER_ORDER_ID" property="partner_order_id" />
+		<result column="PARTNER_USER_ID" property="partner_user_id" />
+		<result column="ITEM_NAME" property="item_name" />
+		<result column="QUANTITY" property="quantity" />
+		<result column="TOTAL_AMOUNT" property="total_amount" />
+		<result column="VAT_AMOUNT" property="vat_amount" />
+		<result column="TAX_FREE_AMOUNT" property="tax_free_amount" />
+		<result column="APPROVAL_URL" property="approval_url" />
+		<result column="FAIL_URL" property="fail_url" />
+		<result column="CANCEL_URL" property="cancel_url" />
+		<result column="PG_TOKEN" property="pg_token" />
+		<result column="PRODUCT_NO" property="productNo" />
 	</resultMap>
+	<insert id="insertParameter">
+		INSERT INTO KAKOREQUESTDTO (
+		C_NO,
+		C_ID,
+		PARTNER_ORDER_ID,
+		PARTNER_USER_ID,
+		ITEM_NAME,
+		QUANTITY,
+		TOTAL_AMOUNT,
+		VAT_AMOUNT,
+		TAX_FREE_AMOUNT,
+		APPROVAL_URL,
+		FAIL_URL,
+		CANCEL_URL
+		)
+		VALUES (
+		SEQ_C_NO.NEXTVAL,
+		#{cid},
+		#{partner_order_id},
+		#{partner_user_id},
+		#{item_name},
+		#{quantity},
+		#{total_amount},
+		#{vat_amount},
+		#{tax_free_amount},
+		#{approval_url},
+		#{fail_url},
+		#{cancel_url}
+		)
+	</insert>
+	<update id="updatePg_token">
+		UPDATE KAKOREQUESTDTO
+		SET PG_TOKEN =	#{pg_token},
+		PRODUCT_NO=#{productNo}
+		WHERE C_NO = (SELECT MAX(C_NO) FROM KAKOREQUESTDTO)
+		
+	</update>
 
-	
 </mapper>
